ontology update
added to BK-Ontology:

bk:TypeAssignment as a subClass of CIDOC E17_TypeAssignment used that a scholar can assign transactions to a specific type (or context) all transactions of the investment project x, all transactions of farm Y).

bk:transactionType as object property linking from Transaction to TypeAssignment

bk:commentary to add meta information to the TypeAssignment

bk:tstatus as data property from transaction to a literal for adding meta information like "pp." or "in full" or "settled" to a transaction.